machine_learning_models/app/requirements.txt
scikit-learn
pandas
numpy
sentence-transformers
joblib
streamlit
matplotlib
seaborn
docx2txt
PyPDF2
python-docx
pdfplumber